requirements.txt
OpenAI
aiohttp
pytest
python-dotenv